Add files via upload
Here is all the codes and models I used in my master thesis. It is important to mention that my Matlab version is R2021a, the version of Prescan is 2019.2.0 and the version of CARSim is 2019.1. The experiment proves this mixture of version works well. If you used other versions, I can not make sure the model and codes could work well.